+++ frontend/src/lib/utils.js
@@ -2,6 +2,7 @@ import React from 'react'
 import api from './api'
 import { toast } from 'react-toastify'
 import PropTypes from 'prop-types'
+import { Spin } from 'antd'
 
 export function uuid() {
     return ([1e7] + -1e3 + -4e3 + -8e3 + -1e11).replace(/[018]/g, c =>
@@ -44,7 +45,16 @@ export let percentage = division =>
 export let Loading = () => (
     <div className="loading-overlay">
         <div></div>
+        <Spin />
     </div>
+)
+
+export const TableRowLoading = ({ colSpan = 1, asOverlay = false }) => (
+    <tr className={asOverlay ? 'loading-overlay over-table' : ''}>
+        <td colSpan={colSpan} style={{ padding: 50, textAlign: 'center' }}>
+            <Spin />
+        </td>
+    </tr>
 )
 
 export let CloseButton = props => {
@@ -166,3 +176,31 @@ export let debounce = (func, wait, immediate) => {
 export const capitalizeFirstLetter = string => {
     return string.charAt(0).toUpperCase() + string.slice(1)
 }
+
+export const operatorMap = {
+    exact: '= equals',
+    is_not: "≠ doesn't equal",
+    icontains: '∋ contains',
+    not_icontains: "∌ doesn't contain",
+    gt: '> greater than',
+    lt: '< lower than',
+}
+
+const operatorEntries = Object.entries(operatorMap).reverse()
+
+export const formatFilterName = str => {
+    for (let [key, value] of operatorEntries) {
+        if (str.includes(key)) return str.replace('__' + key, '') + ` ${value.split(' ')[0]} `
+    }
+    return str + ` ${operatorMap['exact'].split(' ')[0]} `
+}
+
+export const deletePersonData = (person, callback) => {
+    window.confirm('Are you sure you want to delete this user? This cannot be undone') &&
+        api.delete('api/person/' + person.id).then(() => {
+            toast('Person succesfully deleted.')
+            if (callback) callback()
+        })
+}
+
+export const objectsEqual = (obj1, obj2) => JSON.stringify(obj1) === JSON.stringify(obj2)