+++ frontend/src/scenes/actions/ActionsTable.js
@@ -1,14 +1,14 @@
 import React, { Component } from 'react'
-import api from '../../lib/api'
-import { Link } from 'react-router-dom'
-import { DeleteWithUndo } from '../../lib/utils'
-import { AppEditorLink } from '../../lib/components/AppEditorLink/AppEditorLink'
+import api from 'lib/api'
+import { Link } from 'lib/components/Link'
+import { DeleteWithUndo, TableRowLoading } from 'lib/utils'
 
 export class ActionsTable extends Component {
     constructor(props) {
         super(props)
 
         this.state = {
+            actions: [],
             newEvents: [],
             loading: true,
         }
@@ -36,28 +36,18 @@ export class ActionsTable extends Component {
                 <h1>Actions</h1>
                 <p style={{ maxWidth: 600 }}>
                     <i>
-                        Actions are PostHog’s way of easily cleaning up a large
-                        amount of Event data. Actions consist of one or more
-                        events that you have decided to put into a
-                        manually-labelled bucket. They're used in Funnels, Live
-                        actions and Trends.
+                        Actions are PostHog’s way of easily cleaning up a large amount of Event data. Actions consist of
+                        one or more events that you have decided to put into a manually-labelled bucket. They're used in
+                        Funnels, Live actions and Trends.
                         <br />
                         <br />
-                        <a
-                            href="https://github.com/PostHog/posthog/wiki/Actions"
-                            target="_blank"
-                        >
+                        <a href="https://github.com/PostHog/posthog/wiki/Actions" target="_blank">
                             See documentation
                         </a>
                     </i>
                 </p>
 
                 <table className="table" style={{ position: 'relative' }}>
-                    {loading && (
-                        <div className="loading-overlay">
-                            <div></div>
-                        </div>
-                    )}
                     <thead>
                         <tr>
                             <th scope="col">Action ID</th>
@@ -67,7 +57,8 @@ export class ActionsTable extends Component {
                         </tr>
                     </thead>
                     <tbody>
-                        {actions && actions.length == 0 && (
+                        {loading && <TableRowLoading colSpan={4} asOverlay={actions.length > 0} />}
+                        {actions && actions.length === 0 && (
                             <tr>
                                 <td>You don't have any actions yet.</td>
                             </tr>
@@ -76,9 +67,7 @@ export class ActionsTable extends Component {
                             actions.map(action => (
                                 <tr key={action.id}>
                                     <td>
-                                        <Link to={'/action/' + action.id}>
-                                            {action.name}
-                                        </Link>
+                                        <Link to={'/action/' + action.id}>{action.name}</Link>
                                     </td>
                                     <td>{action.count}</td>
                                     <td>
@@ -89,15 +78,9 @@ export class ActionsTable extends Component {
                                                         case '$autocapture':
                                                             return 'Autocapture'
                                                         case '$pageview':
-                                                            return (
-                                                                'URL matches ' +
-                                                                step.url
-                                                            )
+                                                            return 'URL matches ' + step.url
                                                         default:
-                                                            return (
-                                                                'Event: ' +
-                                                                step.event
-                                                            )
+                                                            return 'Event: ' + step.event
                                                     }
                                                 })()}
                                             </div>