+++ webpack.config.js
@@ -4,6 +4,7 @@ const CopyWebpackPlugin = require('copy-webpack-plugin')
 
 module.exports = {
     mode: process.env.NODE_ENV === 'production' ? 'production' : 'development',
+    devtool: process.env.NODE_ENV === 'production' ? 'source-map' : 'inline-source-map',
     entry: {
         main: './frontend/src/index.js',
         editor: './frontend/src/editor/index.js',